Moved serialization from Timeline to Persistence
